requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.14.0
firebase-admin==6.5.0
git-filter-repo==2.45.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
packaging==24.1
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
stevedore==5.2.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
whitenoise==6.7.0
